chore: Expected is in Error.hpp
fix #362
